[ADD] User Security and Role. Turno SemanaProxima [FIX] Turno actualizado.

+++ src/main/java/com/proyectointegrador/clinicaOdontologica/service/impl/OdontologoServiceImpl.java
@@ -2,7 +2,10 @@ package com.proyectointegrador.clinicaOdontologica.service.impl;
 
 import com.proyectointegrador.clinicaOdontologica.dto.OdontologoDTO;
 import com.proyectointegrador.clinicaOdontologica.dto.PacienteDTO;
+import com.proyectointegrador.clinicaOdontologica.exceptions.BadRequestException;
+import com.proyectointegrador.clinicaOdontologica.exceptions.ResourceNotFoundException;
 import com.proyectointegrador.clinicaOdontologica.persistence.entities.Odontologo;
+import com.proyectointegrador.clinicaOdontologica.persistence.entities.Paciente;
 import com.proyectointegrador.clinicaOdontologica.persistence.repositories.IOdontologoRepository;
 import com.proyectointegrador.clinicaOdontologica.service.IService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,8 +27,13 @@ public class OdontologoServiceImpl implements IService<OdontologoDTO> {
     }
 
     @Override
-    public OdontologoDTO buscarPorID(Integer id) {
-         return new OdontologoDTO(odontologoRepository.getById(id));
+    public OdontologoDTO buscarPorID(Integer id) throws ResourceNotFoundException {
+        Odontologo odontologo = odontologoRepository.findById(id).orElse(null);
+        if(odontologo != null) {
+        return new OdontologoDTO(odontologoRepository.getById(id));
+        }else{
+            throw new ResourceNotFoundException("Odontologo con id "+ id+ " no encontrado");
+        }
     }
 
     @Override
@@ -38,15 +46,22 @@ public class OdontologoServiceImpl implements IService<OdontologoDTO> {
     }
 
     @Override
-    public void eliminarPorID(Integer id) {
-        odontologoRepository.deleteById(id);
+    public void eliminarPorID(Integer id) throws ResourceNotFoundException {
+        Odontologo odontologo = odontologoRepository.findById(id).orElse(null);
+        if(odontologo != null) {
+            odontologoRepository.deleteById(id);
+        }else{
+            throw new ResourceNotFoundException("Odontologo con id "+ id+ " no encontrado");
+        }
     }
 
     @Override
-    public OdontologoDTO actualizar(OdontologoDTO o) {
+    public OdontologoDTO actualizar(OdontologoDTO o) throws BadRequestException {
         Odontologo actualizado = null;
         if(odontologoRepository.findById(o.getId()).isPresent()){
             actualizado = odontologoRepository.save(o.toEntity());
+        }else{
+            throw new BadRequestException("odontologo inexistente");
         }
         return new OdontologoDTO(actualizado);
     }
