+++ src/main/java/com/proyectointegrador/clinicaOdontologica/service/impl/PacienteServiceImpl.java
@@ -1,6 +1,8 @@
 package com.proyectointegrador.clinicaOdontologica.service.impl;
 
 import com.proyectointegrador.clinicaOdontologica.dto.PacienteDTO;
+import com.proyectointegrador.clinicaOdontologica.exceptions.BadRequestException;
+import com.proyectointegrador.clinicaOdontologica.exceptions.ResourceNotFoundException;
 import com.proyectointegrador.clinicaOdontologica.persistence.entities.Domicilio;
 import com.proyectointegrador.clinicaOdontologica.persistence.entities.Paciente;
 import com.proyectointegrador.clinicaOdontologica.persistence.repositories.IPacienteRepository;
@@ -25,8 +27,13 @@ public class PacienteServiceImpl implements IService<PacienteDTO> {
     }
 
     @Override
-    public PacienteDTO buscarPorID(Integer id) {
-        return new PacienteDTO(pacienteRepository.getById(id));
+    public PacienteDTO buscarPorID(Integer id) throws ResourceNotFoundException {
+        Paciente paciente = pacienteRepository.findById(id).orElse(null);
+        if(paciente != null) {
+            return new PacienteDTO(pacienteRepository.getById(id));
+        }else{
+            throw new ResourceNotFoundException("Paciente con id "+ id+ " no encontrado");
+        }
     }
 
     @Override
@@ -39,15 +46,22 @@ public class PacienteServiceImpl implements IService<PacienteDTO> {
     }
 
     @Override
-    public void eliminarPorID(Integer id) {
-        pacienteRepository.deleteById(id);
+    public void eliminarPorID(Integer id) throws ResourceNotFoundException {
+        Paciente paciente = pacienteRepository.findById(id).orElse(null);
+        if(paciente != null) {
+            pacienteRepository.deleteById(id);
+        }else{
+            throw new ResourceNotFoundException("Paciente con id "+ id+ " no encontrado");
+        }
     }
 
     @Override
-    public PacienteDTO actualizar(PacienteDTO p) {
+    public PacienteDTO actualizar(PacienteDTO p) throws BadRequestException {
         Paciente actualizado = null;
         if(pacienteRepository.findById(p.getId()).isPresent()){
             actualizado = pacienteRepository.save(p.toEntity());
+        }else{
+            throw new BadRequestException("paciente inexistente");
         }
         return new PacienteDTO(actualizado);
     }
